Try to fix System.IO.IOException  (Cannot create a file when that file already exists.) raised on File.Move reported by Tran Quang Minh

+++ trunk/Translate.Net/source/Libraries/FreeCL.Net/source/rtl/XMLSerializableObject.cs
@@ -113,7 +113,15 @@ namespace FreeCL.RTL
 			writer.Close();
 			writer.Dispose();
 			
-			File.Delete(fileName);
+			for(int i = 0; i < 3; i++)
+			{
+				if(File.Exists(fileName))
+					File.Delete(fileName);
+				else
+					break;
+				if(File.Exists(fileName))
+					System.Threading.Thread.Sleep(250);
+			}
 			try 
 			{
 				File.Move(backupFileName, fileName);
